Windows port: sliding_window: fix MSVC narrowing conversions
- halo/device/untilize_with_halo_program_factory.cpp: static_cast nhw_index
  (size_t) to uint32_t before use in a std::vector<uint32_t> braced init.
- sliding_window.cpp: static_cast src_offset/dst_offset/transfer_size
  (uint32_t) to uint16_t before GatherTransfer braced-init (fields are
  uint16_t).

+++ ttnn/cpp/ttnn/operations/sliding_window/halo/device/untilize_with_halo_program_factory.cpp
@@ -369,7 +369,7 @@ ProgramDescriptor build_halo_program(
                 reader_0_desc.runtime_args.emplace_back(core, std::vector<uint32_t>{0});
                 reader_1_desc.runtime_args.emplace_back(core, std::vector<uint32_t>{0});
             } else if (is_block_sharded) {
-                const auto nhw_index = is_rm_orientation ? core.y : core.x;
+                const auto nhw_index = static_cast<uint32_t>(is_rm_orientation ? core.y : core.x);
                 reader_0_desc.runtime_args.emplace_back(core, std::vector<uint32_t>{nhw_index});
                 reader_1_desc.runtime_args.emplace_back(core, std::vector<uint32_t>{nhw_index});
             }
